refactor: separate shared prose from MDX components
