+++ src/app/(home)/work/[slug]/page.tsx
@@ -5,8 +5,8 @@ import { type ReactElement, ViewTransition } from 'react'
 import { BlurImage } from '@/components/blur-image'
 import { mdxComponents } from '@/components/content/mdx/components'
 import { GitHubCode } from '@/components/content/mdx/github-code'
-import { PreviewButton } from '@/components/content/mdx/preview-button'
 import { TOCMinimap, TOCPopover } from '@/components/features/toc'
+import { PreviewButton } from '@/components/features/work/preview-button'
 import { Icons } from '@/components/icons/icons'
 import { WorkJsonLd } from '@/components/json-ld'
 import { Section } from '@/components/layout/sections'
