+++ src/components/content/mdx/components.tsx
@@ -7,10 +7,10 @@ import { Step, Steps } from 'fumadocs-ui/components/steps'
 import { Tab, Tabs } from 'fumadocs-ui/components/tabs'
 import defaultMdxComponents from 'fumadocs-ui/mdx'
 import type { MDXComponents } from 'mdx/types'
-import { Heading } from '@/components/content/mdx/prose'
 import { VideoPlayer } from '../../ui/video-player'
 import { FramedImage, IframeEmbed } from './embed'
 import { GitHubRepo, GithubRepo } from './github-repo'
+import { Heading } from './heading'
 import { LinkPreviewAnchor } from './link-preview'
 import { Mermaid } from './mermaid'
