+++ src/components/content/mdx/heading.tsx
@@ -1,26 +1,6 @@
-import { Slot } from '@radix-ui/react-slot'
 import type { ComponentProps } from 'react'
 import { Icons } from '@/components/icons/icons'
 import { cn } from '@/lib/utils'
-
-export function Prose({
-  className,
-  asChild = false,
-  ...props
-}: ComponentProps<'div'> & { asChild?: boolean }) {
-  const Comp = asChild ? Slot : 'div'
-
-  return (
-    <Comp
-      className={cn(
-        'prose prose-zinc dark:prose-invert prose-content max-w-none',
-        className
-      )}
-      data-slot='prose'
-      {...props}
-    />
-  )
-}
 
 type HeadingTag = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6'
